Move ZipOptions to types.ts
