+++ src/zip.ts
@@ -2,37 +2,9 @@ import allSettled from 'promise.allsettled';
 import { Writer } from '@transcend-io/conflux';
 import { createWriteStream } from 'streamsaver';
 import mime from 'mime-types';
-import type { PenumbraFile, ZipProgressEmit, ZipCompletionEmit, RemoteResource } from './types';
+import type { PenumbraFile, ZipOptions } from './types';
 import emitZipProgress from './utils/emitZipProgress';
 import emitZipCompletion from './utils/emitZipCompletion';
-
-/** PenumbraZipWriter constructor options */
-export type ZipOptions = Partial<{
-  /** Filename to save to (.zip is optional) */
-  name?: string;
-  /** Total size of archive (if known ahead of time, for 'store' compression level) */
-  size?: number;
-  /** Files (in-memory & remote) to add to zip archive */
-  files: PenumbraFile[];
-  /** Abort controller for cancelling zip generation and saving */
-  controller: AbortController;
-  /** Allow & auto-rename duplicate files sent to writer. Defaults to on */
-  allowDuplicates: boolean;
-  /** Zip archive compression level */
-  compressionLevel: number;
-  /** Store a copy of the resultant zip file in-memory for inspection & testing */
-  saveBuffer: boolean;
-  /**
-   * Auto-registered `'progress'` event listener. This is equivalent to calling
-   * `PenumbraZipWriter.addEventListener('progress', onProgress)`
-   */
-  onProgress?(event: ZipProgressEmit): void;
-  /**
-   * Auto-registered `'write-complete'` event listener. This is equivalent to calling
-   * `PenumbraZipWriter.addEventListener('complete', onComplete)`
-   */
-  onComplete?(event: ZipCompletionEmit): void;
-}>;
 
 /** Compression levels */
 export enum Compression {
